% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binom.R
\name{BH.log}
\alias{BH.log}
\title{Benjamini-Hochberg (FDR) correction of log(p).
code taken from p.adjust and log10 terms added}
\usage{
BH.log(log.p)
}
\description{
Benjamini-Hochberg (FDR) correction of log(p).
code taken from p.adjust and log10 terms added
}
