% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binom.R
\name{binomial}
\alias{binomial}
\title{id - Column name in infile containing ids. By default it is guessed.
count - Column name in infile containing counts. Required.
annotation - Column name containing annotation terms to test enrichment for, e.g. GO. Required.
group - Column name containing grouping label. Default is doing a single ORA for all data.}
\usage{
binomial(annotations, id, count, counts, annotation, group)
}
\description{
id - Column name in infile containing ids. By default it is guessed.
count - Column name in infile containing counts. Required.
annotation - Column name containing annotation terms to test enrichment for, e.g. GO. Required.
group - Column name containing grouping label. Default is doing a single ORA for all data.
}
