% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisher.R
\name{enriched}
\alias{enriched}
\title{get table with columns annotation, p, q for enriched annotation terms.
annotations: table that contains columns id.col, annotation.col
gois: genes of interest. names found in id.col column of annotation table.
id.col: string name of column with ids to match
q.threshold: threshold for q-value signifying significance}
\usage{
enriched(annotations, gois, id.col, annotation.col, q.threshold = 0.05)
}
\description{
get table with columns annotation, p, q for enriched annotation terms.
annotations: table that contains columns id.col, annotation.col
gois: genes of interest. names found in id.col column of annotation table.
id.col: string name of column with ids to match
q.threshold: threshold for q-value signifying significance
}
