% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisher.R
\name{fisher.p}
\alias{fisher.p}
\title{read gene list of interest (GOI) from stdin and write enrichment table to stdout.
You can use merge.R -s and a synonym2object.tsv to convert gene symbols to GO object IDs.
Take a single positional arg with a table of GO annotations, at least columns "GO" and one with the id.
also takes a required positional arg with the column name for ids found in both infile and the GO annotation table.
Can be used for other annotation than specifically GO annotation by setting arg -a/--annotation
fisher's test using x, m, n, k naming like dhyper.
this function is identical to phyper(x-1, m, n, k, lower.tail=F)}
\usage{
fisher.p(x, m, n, k)
}
\description{
read gene list of interest (GOI) from stdin and write enrichment table to stdout.
You can use merge.R -s and a synonym2object.tsv to convert gene symbols to GO object IDs.
Take a single positional arg with a table of GO annotations, at least columns "GO" and one with the id.
also takes a required positional arg with the column name for ids found in both infile and the GO annotation table.
Can be used for other annotation than specifically GO annotation by setting arg -a/--annotation
fisher's test using x, m, n, k naming like dhyper.
this function is identical to phyper(x-1, m, n, k, lower.tail=F)
}
